'fix' tests
