+++ package.json
@@ -75,7 +75,7 @@
     "create-react-class": "^15.5.3",
     "cross-env": "^2.0.0",
     "enzyme": "^3.1.0",
-    "enzyme-adapter-react-15": "^1.0.1",
+    "enzyme-adapter-react-16": "^1.1.1",
     "eslint": "^4.8.0",
     "eslint-config-airbnb": "^15.1.0",
     "eslint-config-prettier": "^2.9.0",
@@ -86,8 +86,8 @@
     "eslint-plugin-react": "^7.5.1",
     "fs-extra": "^4.0.2",
     "husky": "^0.14.3",
-    "karma": "^1.1.1",
-    "karma-chrome-launcher": "^1.0.1",
+    "karma": "^2.0.0",
+    "karma-chrome-launcher": "^2.2.0",
     "karma-cli": "^1.0.1",
     "karma-coverage": "^1.1.0",
     "karma-firefox-launcher": "^1.0.0",
@@ -95,7 +95,7 @@
     "karma-mocha-reporter": "^2.0.4",
     "karma-sinon-chai": "^1.2.2",
     "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^2.0.4",
+    "karma-webpack": "^2.0.13",
     "lint-staged": "^6.0.0",
     "lodash": "^4.13.1",
     "mocha": "^2.5.3",
@@ -106,9 +106,9 @@
     "release-script": "^1.0.2",
     "sinon": "^2.3.2",
     "sinon-chai": "^2.8.0",
-    "webpack": "^1.13.1",
+    "webpack": "^3.0.0",
     "webpack-atoms": "^4.1.2",
-    "webpack-dev-server": "^1.14.1",
+    "webpack-dev-server": "^2.0.0",
     "yargs": "^4.7.1"
   },
   "dependencies": {
