+++ test/index.js
@@ -3,7 +3,7 @@ import deprecated from 'prop-types-extra/lib/deprecated';
 import { _resetWarned } from '../src/utils/deprecationWarning';
 
 import Enzyme, { ShallowWrapper, ReactWrapper } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-15';
+import Adapter from 'enzyme-adapter-react-16';
 
 Enzyme.configure({ adapter: new Adapter() });
